% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_check.R
\name{palette_check}
\alias{palette_check}
\title{Title}
\usage{
palette_check(x, tolerance = NULL, ...)
}
\arguments{
\item{x}{}

\item{tolerance}{}

\item{...}{}
}
\description{
Title
}
\examples{
x = rcartocolor::carto_pal(11, "Vivid")
palette_check(x)
palette_check(x, tolerance = 1)
palette_check(x, tolerance = 10, metric = 2000)
}
