% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_dist.R
\name{palette_dist}
\alias{palette_dist}
\title{Title}
\usage{
palette_dist(x, cvd = NULL, metric = 1976, ...)
}
\arguments{
\item{x}{}

\item{cvd}{}

\item{metric}{a vector of color metric specifiers. Valid values are '1976','1994', and '2000', which refer to the year the metric  was recommended by the CIE.}

\item{...}{}
}
\description{
Title
}
\examples{
x = rcartocolor::carto_pal(11, "Vivid")
palette_dist(x)
palette_dist(x, cvd = "deu")

}
