% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorblindcheck-package.R
\docType{package}
\name{colorblindcheck-package}
\alias{colorblindcheck}
\alias{colorblindcheck-package}
\title{colorblindcheck: Check Color Palletes for Problems with Color Vision Deficiency}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (0000-0002-1057-3721)

}
\keyword{internal}
