% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorblindcheck-package.R
\docType{package}
\name{colorblindcheck-package}
\alias{colorblindcheck}
\alias{colorblindcheck-package}
\title{colorblindcheck: Check Color Palettes for Problems with Color Vision Deficiency}
\description{
Compare color palettes with simulations of color vision deficiencies - deuteranopia, protanopia, and tritanopia.
        It includes calculation of distances between colors, and creating summaries of differences between a color palette and simulations of color vision deficiencies.
        This work was inspired by the blog post at <http://www.vis4.net/blog/2018/02/automate-colorblind-checking/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Nowosad/colorblindcheck}
  \item Report bugs at \url{https://github.com/Nowosad/colorblindcheck/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

}
\keyword{internal}
