% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_plot.R
\name{palette_plot}
\alias{palette_plot}
\title{Title}
\usage{
palette_plot(x)
}
\arguments{
\item{x}{}
}
\description{
Title
}
\examples{
palette_plot(x = rcartocolor::carto_pal(7, "Sunset"))
palette_plot(x = rcartocolor::carto_pal(11, "Safe"))
palette_plot(x = rcartocolor::carto_pal(7, "Earth"))
palette_plot(x = rcartocolor::carto_pal(11, "Vivid"))
}
